Add ::deep combinator to style active tree nodes in shadow DOM
The CSS selector for the `.tree-node.is-active` class within the
`.menu.rf-treeview` class has been modified. The `::deep` combinator
has been added to the `.tree-node.is-active a` selector to ensure
that styles are applied deeply within the shadow DOM. This change
ensures that both `button` and `a` elements within an active tree
node are styled correctly.

+semver: patch